docs/old/googleanalytics2/requirements.txt
alembic==1.12.0
blinker==1.9.0
build==1.2.2.post1
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
cryptography==42.0.5
Flask==2.3.3
Flask-Migrate==4.0.5
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.0
google-analytics-admin>=0.6.0,<0.7.0
google-analytics-data==0.16.2
google-api-core==2.25.0rc1
google-api-python-client==2.100.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.62.3
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.9.4
msgspec==0.19.0
narwhals==1.38.2
numpy==2.2.5
oauthlib==3.2.2
packaging==25.0
pandas==2.2.3
pdfkit==1.0.0
pillow==11.2.1
pip-tools==7.4.1
playwright==1.52.0
plotly==6.0.1
proto-plus==1.26.1
protobuf==4.25.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyee==13.0.0
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.4.0
six==1.17.0
SQLAlchemy==2.0.21
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.20
Werkzeug==3.0.6
wheel==0.45.1
